Add cache interface
